fix(gguf): fix NaN logits from meta RoPE buffers and FP16 overflow
Three bugs in the GGUF→HuggingFace loading pipeline:

1. RoPE inv_freq buffers stayed as meta tensors after load_state_dict
   with assign=True + strict=False, causing "Cannot copy out of meta
   tensor" errors. Fix: reinitialize all LlamaRotaryEmbedding modules
   after loading (both model-level and per-layer).

2. nan_to_num converted ±Inf to ±FLT_MAX instead of zero in Q4_K/Q5_K/
   Q6_K dequantization. Padding blocks can produce Inf scales, not just
   NaN. Fix: nan_to_num(nan=0, posinf=0, neginf=0).

3. Float32 dequantized values overflowed FP16 range (±65504) when
   converted via .to(float16), producing Inf weights in lm_head.
   Fix: clamp to dtype range before conversion, and use BF16 (same
   exponent range as FP32) as the default GGUF dequant dtype for
   NF4/INT8 modes.

+++ llm_surgeon/surgery.py
@@ -576,7 +576,7 @@ def load_model(model_id: str, mode: str = "nf4") -> Tuple:
         blob = resolve_ollama_blob(model_id)
         if blob is not None:
             _GGUF_DTYPE = {
-                "nf4": torch.float16, "int8": torch.float16,
+                "nf4": torch.bfloat16, "int8": torch.bfloat16,
                 "bf16": torch.bfloat16, "fp16": torch.float16,
                 "fp32": torch.float32, "fp32-cpu": torch.float32,
             }
